Fix in JSON files not having write permissions
NOTE: Terrible fix. WE SHOULD NOT BE USING JSON FILES FOR REGISTERING INFORMATION IN REACT NATIVE. This fix forces the android to create the json file themselves with the previous information. Which in itself actually shouldn't be a problem because we SHOULD NOT be parsing previous information of the favorites to the android system. Why are we reading a previous file?????????? We should only be writing in it :facepalm